src/embedding/lua_embedder.py
import os
import re
from typing import Dict, List, Tuple, Any, Optional

import numpy as np
from tree_sitter import Language, Parser

# This is a placeholder for actual embedding generation
# In the real implementation, this would use a model like OpenAI or a local alternative
def generate_embedding(text: str) -> List[float]:
    """
    Generate a vector embedding for a text snippet.
    This is a placeholder that creates a random vector.
    """
    # In a real implementation, this would call an embedding model
    # For now, we return a random vector of size 1536 (typical OpenAI size)
    return np.random.randn(1536).tolist()

class LuaEmbedder:
    def __init__(self, tree_sitter_lib_path: Optional[str] = None):
        """
        Initialize the Lua embedder with tree-sitter.
        
        Args:
            tree_sitter_lib_path: Path to the tree-sitter library
        """
        # In a real implementation, we would:
        # 1. Load the tree-sitter Lua grammar
        # 2. Set up the parser
        self.parser = None
        self.language = None
        
        # This is just a placeholder for the implementation
        print("Initializing Lua embedder (placeholder)")
        
    def parse_file(self, file_path: str) -> List[Dict[str, Any]]:
        """
        Parse a Lua file and extract code chunks.
        
        Args:
            file_path: Path to the Lua file
            
        Returns:
            List of code chunks with metadata
        """
        # In a real implementation, we would:
        # 1. Read the file
        # 2. Parse it with tree-sitter
        # 3. Extract meaningful chunks based on the AST
        # 4. Generate embeddings for each chunk
        
        # This is just a placeholder that simulates reading the file
        try:
            with open(file_path, 'r', encoding='utf-8') as f:
                content = f.read()
        except Exception as e:
            print(f"Error reading file {file_path}: {e}")
            return []
            
        # Simple chunking by function definitions (placeholder)
        chunks = []
        lines = content.split('\n')
        
        # Very simple regex pattern to find function definitions
        function_pattern = re.compile(r'function\s+(\w+(?:\.\w+)*(?:\:\w+)?)?\s*\(')
        
        chunk_start = 0
        in_function = False
        current_function = None
        
        for i, line in enumerate(lines):
            # Check for function start
            if not in_function:
                match = function_pattern.search(line)
                if match:
                    in_function = True
                    current_function = match.group(1) or "anonymous"
                    chunk_start = i
                    
            # Check for function end
            elif "end" in line and not any(keyword in line for keyword in ["if", "for", "while"]):
                # This is a very simplistic approach, would need proper parsing in real implementation
                chunk_content = "\n".join(lines[chunk_start:i+1])
                
                chunks.append({
                    "chunk_type": "function",
                    "start_line": chunk_start + 1,  # 1-indexed for DB
                    "end_line": i + 1,  # 1-indexed for DB
                    "content": chunk_content,
                    "embedding": generate_embedding(chunk_content),
                    "metadata": {
                        "function_name": current_function
                    }
                })
                
                in_function = False
                current_function = None
                
        # Add a chunk for the whole file
        chunks.append({
            "chunk_type": "file",
            "start_line": 1,
            "end_line": len(lines),
            "content": content,
            "embedding": generate_embedding(content),
            "metadata": {
                "file_path": file_path,
                "file_name": os.path.basename(file_path)
            }
        })
        
        return chunks