src/embedding/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.4.2
asyncpg==0.28.0
httpx==0.25.1
numpy==1.26.1
pgvector==0.2.3
llama-index==0.9.3
transformers==4.35.0
torch==2.1.0
tokenizers==0.14.1
lua-ast==0.1.7
sentence-transformers==2.2.2
langchain==0.0.325
tenacity==8.2.3